added inv_id to inv_table

+++ database/migrations/2019_11_09_000012_add_relationship_fields_to_comments_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToCommentsTable extends Migration
     public function up()
     {
         Schema::table('comments', function (Blueprint $table) {
-            $table->unsignedInteger('ticket_id')->nullable();
+            $table->unsignedBigInteger('ticket_id')->nullable();
 
             $table->foreign('ticket_id', 'ticket_fk_583774')->references('id')->on('tickets');
